requirements.txt
azure-storage-blob
pandas
numpy
matplotlib
seaborn
tqdm
scikit-learn
joblib
python-dotenv
nltk
gensim
imblearn
streamlit
psycopg2
sqlalchemy
jupyter
spacy
pytest
xgboost
